Refactor logic for obtaining AWS credentials
